Added:API Routes Separation #12
Added api routes sparation in a routes dir and moved the sample database
data into a db folder in a data.py and imported the db data into
specific routes files.
