+++ routes/about.py
@@ -1,0 +1,18 @@
+from fastapi import Request, APIRouter
+from fastapi.responses import HTMLResponse
+from fastapi.templating import Jinja2Templates
+from db.data import team_members
+
+router = APIRouter()
+templates = Jinja2Templates(directory="templates")
+
+@router.get("/about", response_class=HTMLResponse)
+async def about(request: Request):
+    """About page with community information"""
+    context = {
+        "request": request,
+        "title": "About - CNCF Kathmandu",
+        "team_members": team_members,
+        "description": "We are a community of cloud native enthusiasts in Kathmandu, Nepal. Our mission is to promote cloud native technologies and help developers learn and grow together."
+    }
+    return templates.TemplateResponse("about.html", context)
