+++ routes/home.py
@@ -1,0 +1,21 @@
+from fastapi import Request, APIRouter
+from fastapi.responses import HTMLResponse
+from fastapi.templating import Jinja2Templates
+from db.data import events_db
+
+router = APIRouter()
+templates = Jinja2Templates(directory="templates")
+
+@router.get("/", response_class=HTMLResponse)
+async def home(request: Request):
+    """Home page displaying community information"""
+    context = {
+        "request": request,
+        "title": "CNCF Kathmandu - Home",
+        "community_name": "CNCF Kathmandu",
+        "tagline": "Building the Future of Cloud Native Computing",
+        "upcoming_events": [e for e in events_db if e["status"] == "upcoming"][:3]
+    }
+    return templates.TemplateResponse("index.html", context)
+
+
